# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Fund Analysis API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/data/jobs/recent": {
      "get": {
        "tags": [
          "data"
        ],
        "summary": "Recent Jobs",
        "operationId": "recent_jobs_api_data_jobs_recent_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 50,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/data/sync": {
      "post": {
        "tags": [
          "data"
        ],
        "summary": "Sync Data",
        "operationId": "sync_data_api_data_sync_post",
        "parameters": [
          {
            "name": "task",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "all",
                "profiles",
                "navs",
                "metrics",
                "risk_levels"
              ],
              "type": "string",
              "default": "all",
              "title": "Task"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/funds/search": {
      "get": {
        "tags": [
          "funds"
        ],
        "summary": "Search",
        "operationId": "search_api_funds_search_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          },
          {
            "name": "fund_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "enum": [
                      "stock",
                      "mixed",
                      "bond",
                      "money"
                    ],
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Fund Type"
            }
          },
          {
            "name": "risk_level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "array",
                  "items": {
                    "enum": [
                      "R1",
                      "R2",
                      "R3",
                      "R4",
                      "R5"
                    ],
                    "type": "string"
                  }
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risk Level"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "code",
                "annualized_return_3y",
                "annualized_return_5y",
                "max_drawdown",
                "sharpe_ratio",
                "fee",
                "size"
              ],
              "type": "string",
              "default": "code",
              "title": "Sort By"
            }
          },
          {
            "name": "sort_order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "asc",
              "title": "Sort Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/funds/filter": {
      "post": {
        "tags": [
          "funds"
        ],
        "summary": "Filter Endpoint",
        "operationId": "filter_endpoint_api_funds_filter_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FundFilterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/funds/compare": {
      "post": {
        "tags": [
          "funds"
        ],
        "summary": "Compare",
        "operationId": "compare_api_funds_compare_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FundCompareRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/funds/{code}": {
      "get": {
        "tags": [
          "funds"
        ],
        "summary": "Detail",
        "operationId": "detail_api_funds__code__get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "risk_profile",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "C1",
                "C2",
                "C3",
                "C4",
                "C5"
              ],
              "type": "string",
              "default": "C3",
              "title": "Risk Profile"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/funds/{code}/nav": {
      "get": {
        "tags": [
          "funds"
        ],
        "summary": "Nav",
        "operationId": "nav_api_funds__code__nav_get",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Code"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Start"
            }
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "End"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/risk-assessments": {
      "post": {
        "tags": [
          "risk"
        ],
        "summary": "Submit",
        "operationId": "submit_api_risk_assessments_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskAssessmentSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/portfolios": {
      "get": {
        "tags": [
          "portfolios"
        ],
        "summary": "List Portfolios",
        "operationId": "list_portfolios_api_portfolios_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "portfolios"
        ],
        "summary": "Create Portfolio",
        "operationId": "create_portfolio_api_portfolios_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/backtests": {
      "post": {
        "tags": [
          "backtests"
        ],
        "summary": "Create",
        "operationId": "create_api_backtests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BacktestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BacktestRequest": {
        "properties": {
          "strategy_type": {
            "type": "string",
            "enum": [
              "monthly_dca",
              "weekly_dca",
              "rebalance",
              "template_portfolio"
            ],
            "title": "Strategy Type"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "start": {
            "type": "string",
            "title": "Start"
          },
          "end": {
            "type": "string",
            "title": "End"
          },
          "fund_codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 5,
            "minItems": 1,
            "title": "Fund Codes"
          },
          "rebalance_threshold": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Rebalance Threshold",
            "default": 5
          },
          "template_key": {
            "type": "string",
            "enum": [
              "conservative",
              "balanced",
              "growth"
            ],
            "title": "Template Key",
            "default": "balanced"
          }
        },
        "type": "object",
        "required": [
          "strategy_type",
          "amount",
          "start",
          "end"
        ],
        "title": "BacktestRequest"
      },
...
      "FundCompareRequest": {
        "properties": {
          "codes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 5,
            "minItems": 1,
            "title": "Codes"
          }
        },
        "type": "object",
        "required": [
          "codes"
        ],
        "title": "FundCompareRequest"
      },
      "FundFilterRequest": {
        "properties": {
          "risk_profile": {
            "type": "string",
            "enum": [
              "C1",
              "C2",
              "C3",
              "C4",
              "C5"
            ],
            "title": "Risk Profile"
          },
          "fund_types": {
            "items": {
              "type": "string",
              "enum": [
                "stock",
                "mixed",
                "bond",
                "money"
              ]
            },
            "type": "array",
            "title": "Fund Types"
          },
          "min_years": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Min Years"
          },
          "size_range": {
            "prefixItems": [
              {
                "type": "number"
              },
              {
                "type": "number"
              }
            ],
            "type": "array",
            "maxItems": 2,
            "minItems": 2,
            "title": "Size Range"
          },
          "return_rank_percentile": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Return Rank Percentile"
          },
          "max_drawdown_lte_category_avg": {
            "type": "boolean",
            "title": "Max Drawdown Lte Category Avg"
          },
          "sharpe_gte": {
            "type": "number",
            "title": "Sharpe Gte"
          },
          "fee_lte": {
            "type": "number",
            "title": "Fee Lte"
          },
          "sort_by": {
            "type": "string",
            "enum": [
              "code",
              "annualized_return_3y",
              "annualized_return_5y",
              "max_drawdown",
              "sharpe_ratio",
              "fee",
              "size"
            ],
            "title": "Sort By",
            "default": "annualized_return_3y"
          },
          "sort_order": {
            "type": "string",
            "enum": [
              "asc",
              "desc"
            ],
            "title": "Sort Order",
            "default": "desc"
          }
        },
        "type": "object",
        "required": [
          "risk_profile",
          "fund_types",
          "min_years",
          "size_range",
          "return_rank_percentile",
          "max_drawdown_lte_category_avg",
          "sharpe_gte",
          "fee_lte"
        ],
        "title": "FundFilterRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PortfolioCreateRequest": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100,
            "minLength": 1,
            "title": "Name"
          },
          "template_key": {
            "type": "string",
            "enum": [
              "conservative",
              "balanced",
              "growth"
            ],
            "title": "Template Key"
          }
        },
        "type": "object",
        "required": [
          "name",
          "template_key"
        ],
        "title": "PortfolioCreateRequest"
      },
...
      "RiskAnswer": {
        "properties": {
          "question_id": {
            "type": "string",
            "title": "Question Id"
          },
          "score": {
            "type": "integer",
            "maximum": 5.0,
            "minimum": 1.0,
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "question_id",
          "score"
        ],
        "title": "RiskAnswer"
      },
      "RiskAssessmentSubmission": {
        "properties": {
          "answers": {
            "items": {
              "$ref": "#/components/schemas/RiskAnswer"
            },
            "type": "array",
            "maxItems": 8,
            "minItems": 5,
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "answers"
        ],
        "title": "RiskAssessmentSubmission"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}